\name{MEAN_all}
\alias{MEAN_all}
\docType{data}
\title{Mean used for standardizing predictor data in prediction function}
\description{
Column means for all predictor data of original studies used to build the meta-analysis model, used for standardization.
}
\format{
  The format is:
 Named num [1:251] 0.020035 0.002534 0.034492 0.007793 0.000724 ...
 - attr(*, "names")= chr [1:251] "Edge_Type_1" "Edge_Type_2" "Edge_Type_3" "Edge_Type_4" ...
}

\keyword{datasets}
\keyword{internal}
