\name{SD}
\alias{SD}
\docType{data}
\title{Standard deviation used for standardizing predictor data}
\description{
Column standard deviation values for numeric predictor data of original studies used to build the meta-analysis model, used for standardization.
}
\format{
  The format is:
 Named num [1:21] 92.493 1.778 0.559 1.033 0.891 ...
 - attr(*, "names")= chr [1:21] "angle" "IT_Width" "Hab_W1" "Hab_W2" ...
}

\keyword{datasets}
\keyword{internal}
